[synnax] - Homebrew tap version bumping workflow
This commit addresses issue #229 by introducing a GitHub Action
workflow file that automatically bumps the version of synnax
in the homebrew tap. It uses a wrapper for the
`brew bump-formula-pr` cli command and requires a
**non-default** GitHub access token to access the separate
homebrew tap reposity.
The action secret `HOMEBREW_BUMP_TOKEN` needs to be added
in order for this workflow to work.