SY-3110: Fix TypeScript circular dependencies and add React ESLint rules (#1551)
This PR is an attempt at fixing issues that may have caused the React Minified Error 130. I only have medium confidence that these fixes are related, but spent a lot of time trying to replicate and failed. Most of the stuff online that I saw for this error had to do with import / export issues, so I have added a circular dependency check to CI and fixed circular dependencies to attempt to fix this error.

Summary:

- Add a few ESLint rules (react/jsx-no-undef, react/jsx-no-constructed-context-values, react/jsx-no-useless-fragment) that ChatGPT recommended but definitely do not address this issue
- Move ESLint configuration file to TypeScript to take some advantage of that typing, which necessitated moving a few random files to TypeScript
- Fixed circular dependencies in our TypeScript codebase and added a circular dependency check to CI.