SY-3052: Pass Context and Transaction to Gorp Where/Guard/Change Callbacks
There are certain cases, such as in filtering by range labels, where we query nested data structures (i.e. the labels of a range), these queries are currently using the base transaction and a context.Background() or context.TODO() call, which is bad.

Created a new gorp.Context struct which wraps a context and a transaction and passes it to relevant callbacks (Filter, Guard, Change) so that we're using the proper context and transaction.