SY-3152: Fix Synchronization Issues with Workspaces in Console (#1572)
Fix some issues related to workspaces not being saved when switching between Cores in the Console. This includes adding a check when switching Cores to set the current workspace to null, meaning a workspace won't be carried over between Cores. I also removed code that keeps a record of workspaces in the Redux store which removes potential issues with those workspaces and the workspaces on the Core getting out-of-sync.